+++ packages/common/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/common
 
+## 0.9.94
+
+### Patch Changes
+
+- version bump
+
+## 0.9.93
+
+### Patch Changes
+
+- bump
+
+## 0.9.92
+
+### Patch Changes
+
+- bump
+
 ## 0.9.91
 
 ### Patch Changes