+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,27 @@
 # web-ui-sdk
 
+## 1.5.95
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.94
+
+## 1.5.94
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.93
+
+## 1.5.93
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.92
+
 ## 1.5.92
 
 ### Patch Changes