+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.117
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.117
+
+## 0.6.116
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.116
+
+## 0.6.115
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.115
+
 ## 0.6.114
 
 ### Patch Changes