+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @ballerine/workflows-service
 
+## 0.7.124
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.117
+  - @ballerine/common@0.9.94
+  - @ballerine/workflow-node-sdk@0.6.117
+
+## 0.7.123
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.93
+  - @ballerine/workflow-core@0.6.116
+  - @ballerine/workflow-node-sdk@0.6.116
+
+## 0.7.122
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.92
+  - @ballerine/workflow-core@0.6.115
+  - @ballerine/workflow-node-sdk@0.6.115
+
 ## 0.7.121
 
 ### Patch Changes